chor: reorg validation and defaulting constraints implementations
